move data to pages
