+++ frontend/src/lib/components/UserLogin.svelte
@@ -40,22 +40,7 @@
 	}
 
 	// data and variables
-	let data = [
-		{
-			type: 'text',
-			name: 'Benutzerkennung',
-			placeholder: 'Benutzerkennung',
-			label: 'Benutzerkennung',
-			required: true
-		},
-		{
-			type: 'password',
-			name: 'Passwort',
-			placeholder: 'Passwort',
-			label: 'Passwort',
-			required: true
-		}
-	];
+	export let data;
 
 	const buttons = [
 		{
@@ -150,7 +135,7 @@
 		<span class="container mx-auto w-full text-gray-700 dark:text-gray-400">Not registered?</span>
 		<a
 			href={`${base}/userLand/userRegistration`}
-			class="text-primary-700 hover:underline dark:text-primary-500"
+			class="text-primary-700 dark:text-primary-500 hover:underline"
 		>
 			Create account
 		</a>
